+++ data/statistics_around_the_clock.go
@@ -10,9 +10,13 @@ func GetAroundTheClockStatistics(from string, to string) ([]*models.StatisticsAr
 	rows, err := models.DB.Query(`
 		SELECT
 			p.id,
+			COUNT(DISTINCT m.id) AS 'matches_played',
+			COUNT(DISTINCT m2.id) AS 'matches_won',
+			COUNT(DISTINCT l.id) AS 'legs_played',
+			COUNT(DISTINCT l2.id) AS 'legs_won',
 			SUM(s.darts_thrown) as 'darts_thrown',
 			CAST(SUM(s.score) / COUNT(DISTINCT l.id) AS SIGNED) as 'avg_score',
-			CAST(SUM(s.longest_streak) / COUNT(l.id) AS SIGNED) as 'avg_longest_streak',
+			MAX(s.longest_streak) as 'longest_streak',
 			SUM(s.total_hit_rate) / COUNT(l.id) as 'total_hit_rate',
 			SUM(s.hit_rate_1) / COUNT(l.id) as 'hit_rate_1',
 			SUM(s.hit_rate_2) / COUNT(l.id) as 'hit_rate_2',
@@ -55,10 +59,9 @@ func GetAroundTheClockStatistics(from string, to string) ([]*models.StatisticsAr
 	for rows.Next() {
 		s := new(models.StatisticsAroundThe)
 		h := make([]*float32, 26)
-		err := rows.Scan(&s.PlayerID, &s.DartsThrown, &s.Score, &s.LongestStreak, &s.TotalHitRate,
-			&h[1], &h[2], &h[3], &h[4], &h[5], &h[6], &h[7], &h[8], &h[9],
-			&h[10], &h[11], &h[12], &h[13], &h[14], &h[15], &h[16], &h[17],
-			&h[18], &h[19], &h[20], &h[25])
+		err := rows.Scan(&s.PlayerID, &s.MatchesPlayed, &s.MatchesWon, &s.LegsPlayed, &s.LegsWon, &s.DartsThrown,
+			&s.Score, &s.LongestStreak, &s.TotalHitRate, &h[1], &h[2], &h[3], &h[4], &h[5], &h[6], &h[7], &h[8], &h[9],
+			&h[10], &h[11], &h[12], &h[13], &h[14], &h[15], &h[16], &h[17], &h[18], &h[19], &h[20], &h[25])
 		if err != nil {
 			return nil, err
 		}