+++ data/statistics_around_the_world.go
@@ -10,6 +10,10 @@ func GetAroundTheWorldStatistics(from string, to string) ([]*models.StatisticsAr
 	rows, err := models.DB.Query(`
 		SELECT
 			p.id,
+			COUNT(DISTINCT m.id) AS 'matches_played',
+			COUNT(DISTINCT m2.id) AS 'matches_won',
+			COUNT(DISTINCT l.id) AS 'legs_played',
+			COUNT(DISTINCT l2.id) AS 'legs_won',
 			SUM(s.darts_thrown) as 'darts_thrown',
 			CAST(SUM(s.score) / COUNT(DISTINCT l.id) AS SIGNED) as 'avg_score',
 			SUM(s.mpr) / (SUM(s.darts_thrown) / 3) as 'mpr',
@@ -55,9 +59,9 @@ func GetAroundTheWorldStatistics(from string, to string) ([]*models.StatisticsAr
 	for rows.Next() {
 		s := new(models.StatisticsAroundThe)
 		h := make([]*float32, 26)
-		err := rows.Scan(&s.PlayerID, &s.DartsThrown, &s.Score, &s.MPR, &s.TotalHitRate,
-			&h[1], &h[2], &h[3], &h[4], &h[5], &h[6], &h[7], &h[8], &h[9], &h[10], &h[11],
-			&h[12], &h[13], &h[14], &h[15], &h[16], &h[17], &h[18], &h[19], &h[20], &h[25])
+		err := rows.Scan(&s.PlayerID, &s.MatchesPlayed, &s.MatchesWon, &s.LegsPlayed, &s.LegsWon, &s.DartsThrown,
+			&s.Score, &s.MPR, &s.TotalHitRate, &h[1], &h[2], &h[3], &h[4], &h[5], &h[6], &h[7], &h[8], &h[9], &h[10],
+			&h[11], &h[12], &h[13], &h[14], &h[15], &h[16], &h[17], &h[18], &h[19], &h[20], &h[25])
 		if err != nil {
 			return nil, err
 		}
@@ -203,6 +207,10 @@ func GetShanghaiStatistics(from string, to string) ([]*models.StatisticsAroundTh
 	rows, err := models.DB.Query(`
 		SELECT
 			p.id,
+			COUNT(DISTINCT m.id) AS 'matches_played',
+			COUNT(DISTINCT m2.id) AS 'matches_won',
+			COUNT(DISTINCT l.id) AS 'legs_played',
+			COUNT(DISTINCT l2.id) AS 'legs_won',
 			SUM(s.darts_thrown) as 'darts_thrown',
 			CAST(SUM(s.score) / COUNT(DISTINCT l.id) AS SIGNED) as 'avg_score',
 			SUM(s.mpr) / (SUM(s.darts_thrown) / 3) as 'mpr',
@@ -247,9 +255,9 @@ func GetShanghaiStatistics(from string, to string) ([]*models.StatisticsAroundTh
 	for rows.Next() {
 		s := new(models.StatisticsAroundThe)
 		h := make([]*float32, 21)
-		err := rows.Scan(&s.PlayerID, &s.DartsThrown, &s.Score, &s.MPR, &s.TotalHitRate,
-			&h[1], &h[2], &h[3], &h[4], &h[5], &h[6], &h[7], &h[8], &h[9], &h[10], &h[11],
-			&h[12], &h[13], &h[14], &h[15], &h[16], &h[17], &h[18], &h[19], &h[20])
+		err := rows.Scan(&s.PlayerID, &s.MatchesPlayed, &s.MatchesWon, &s.LegsPlayed, &s.LegsWon, &s.DartsThrown,
+			&s.Score, &s.MPR, &s.TotalHitRate, &h[1], &h[2], &h[3], &h[4], &h[5], &h[6], &h[7], &h[8], &h[9], &h[10],
+			&h[11], &h[12], &h[13], &h[14], &h[15], &h[16], &h[17], &h[18], &h[19], &h[20])
 		if err != nil {
 			return nil, err
 		}