+++ data/statistics_shootout.go
@@ -9,7 +9,10 @@ func GetShootoutStatistics(from string, to string) ([]*models.StatisticsShootout
 	rows, err := models.DB.Query(`
 		SELECT
 			p.id AS 'player_id',
-			COUNT(DISTINCT m.id),
+			COUNT(DISTINCT m.id) AS 'matches_played',
+			COUNT(DISTINCT m2.id) AS 'matches_won',
+			COUNT(DISTINCT l.id) AS 'legs_played',
+			COUNT(DISTINCT l2.id) AS 'legs_won',
 			SUM(s.ppd) / COUNT(p.id) AS 'ppd',
 			SUM(s.60s_plus),
 			SUM(s.100s_plus),
@@ -19,54 +22,27 @@ func GetShootoutStatistics(from string, to string) ([]*models.StatisticsShootout
 			JOIN player p ON p.id = s.player_id
 			JOIN leg l ON l.id = s.leg_id
 			JOIN matches m ON m.id = l.match_id
+			LEFT JOIN leg l2 ON l2.id = s.leg_id AND l2.winner_id = p.id
+			LEFT JOIN matches m2 ON m2.id = l.match_id AND m2.winner_id = p.id
 		WHERE m.updated_at >= ? AND m.updated_at < ?
 			AND m.is_finished = 1 AND m.is_abandoned = 0
 			AND m.match_type_id = 2
-		GROUP BY p.id`, from, to)
+		GROUP BY p.id
+		ORDER BY(COUNT(DISTINCT m2.id) / COUNT(DISTINCT m.id)) DESC, matches_played DESC, ppd DESC`, from, to)
 	if err != nil {
 		return nil, err
 	}
 	defer rows.Close()
-
-	statsMap := make(map[int]*models.StatisticsShootout, 0)
-	for rows.Next() {
-		s := new(models.StatisticsShootout)
-		err := rows.Scan(&s.PlayerID, &s.MatchesPlayed, &s.PPD, &s.Score60sPlus, &s.Score100sPlus, &s.Score140sPlus, &s.Score180s)
-		if err != nil {
-			return nil, err
-		}
-		statsMap[s.PlayerID] = s
-	}
-
-	rows, err = models.DB.Query(`
-		SELECT
-			p.id AS 'player_id',
-			COUNT(m.winner_id) AS 'matches_won'
-		FROM matches m
-			JOIN player p ON p.id = m.winner_id
-		WHERE m.updated_at >= ? AND m.updated_at < ?
-		AND m.match_type_id = 2
-		GROUP BY m.winner_id`, from, to)
-	if err != nil {
-		return nil, err
-	}
-	defer rows.Close()
-
-	for rows.Next() {
-		var playerID int
-		var matchesWon int
-		err := rows.Scan(&playerID, &matchesWon)
-		if err != nil {
-			return nil, err
-		}
-		statsMap[playerID].MatchesWon = matchesWon
-	}
 
 	stats := make([]*models.StatisticsShootout, 0)
-	for _, s := range statsMap {
+	for rows.Next() {
+		s := new(models.StatisticsShootout)
+		err := rows.Scan(&s.PlayerID, &s.MatchesPlayed, &s.MatchesWon, &s.LegsPlayed, &s.LegsWon, &s.PPD, &s.Score60sPlus, &s.Score100sPlus, &s.Score140sPlus, &s.Score180s)
+		if err != nil {
+			return nil, err
+		}
 		stats = append(stats, s)
 	}
-
 	return stats, nil
 }