+++ models/statistics_shootout.go
@@ -5,12 +5,14 @@ type StatisticsShootout struct {
 	ID            int             `json:"id,omitempty"`
 	LegID         int             `json:"leg_id,omitempty"`
 	PlayerID      int             `json:"player_id,omitempty"`
+	MatchesPlayed int             `json:"matches_played,omitempty"`
+	MatchesWon    int             `json:"matches_won,omitempty"`
+	LegsPlayed    int             `json:"legs_played,omitempty"`
+	LegsWon       int             `json:"legs_won,omitempty"`
 	PPD           float32         `json:"ppd"`
 	Score60sPlus  int             `json:"scores_60s_plus"`
 	Score100sPlus int             `json:"scores_100s_plus"`
 	Score140sPlus int             `json:"scores_140s_plus"`
 	Score180s     int             `json:"scores_180s"`
-	MatchesPlayed int             `json:"matches_played,omitempty"`
-	MatchesWon    int             `json:"matches_won,omitempty"`
 	Hits          map[int64]*Hits `json:"hits,omitempty"`
 }