+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 #### Feature
 - Support for new game types `Around The World`, `Around The Clock` and `Shanghai`
 - Added match statistics for `9 Dart Shootout`
+- New convenience endpoints`/statistics/<type_id>/<from>/<to>`
+
+#### Fixed
+- Correctly calculate legs played and won for different statistics
+- Fixed calculating of matches and legs played and won for shootout
 
 ## [1.0.0] - 2020-05-03
 #### Feature